feat(luminork): Add support for showing sources on the component GET requests
