+++ lib/luminork-server/src/api_types.rs
@@ -1,5 +1,4 @@
 pub mod actions;
 pub mod change_sets;
-pub mod component;
 pub mod func_run;
 pub mod secrets;
