--- lib/luminork-server/src/api_types/component.rs
@@ -1,1 +1,0 @@
-pub mod v1;
